+++ boundlexx/boundless/models/shop.py
@@ -8,6 +8,7 @@ from django.utils import timezone
 from django_prometheus.models import ExportModelOperationsMixin
 
 from boundlexx.boundless.client import Location, ShopItem
+from boundlexx.boundless.client import World as SimpleWorld
 from boundlexx.boundless.models.game import Item
 from boundlexx.boundless.models.world import World
 from boundlexx.boundless.utils import html_name
@@ -15,7 +16,7 @@ from boundlexx.boundless.utils import html_name
 
 class ItemShopPriceManager(models.Manager):
     def create_from_shop_item(
-        self, world: str, item: Item, shop_item: ShopItem, colors=None
+        self, world: SimpleWorld, item: Item, shop_item: ShopItem, colors=None
     ) -> ItemShopPrice:
         return self.create(
             item_id=item.id,
@@ -31,7 +32,7 @@ class ItemShopPriceManager(models.Manager):
             location_x=shop_item.location.x,
             location_y=shop_item.location.y,
             location_z=shop_item.location.z,
-            world=World.objects.get(name=world, active=True),
+            world=World.objects.get(id=world.id, active=True),
         )